+++ client/src/game/models/templates.ts
@@ -12,7 +12,6 @@ export const BaseTemplateStrings = [
     "stroke_width",
     "name",
     "name_visible",
-    "is_token",
     "default_edit_access",
     "default_movement_access",
     "default_vision_access",